Updated code for dup2, pipe and fork

+++ libc/pipe.c
@@ -1,7 +1,8 @@
 #include <sys/defs.h>
 
 int pipe(int pipefd[2]) {
-	ssize_t output;
+	int64_t output;
+
     __asm__ (
         "movq $22, %%rax;"
         "movq %1, %%rdi;"
